Rename NoneLocation to MemoryLocation
