+++ locations/memory.js
